Add buffer overrun detection to ./stream
When working on lower-capacity devices, it becomes more important to understand
when the inference time is taking long enough that the inbound audio is
overwriting samples that the inference process hasn't yet had a chance to
process.

This patch adds a basic detection scheme to the audio class which maintains a
record of where the client has read up to in the buffer.  If the async process
writes a span into the buffer which overlaps that cursor, it sets a flag that the
client can use to detect whether an overrun has happened.

This mechanism is then used in examples/stream/stream.cpp to signal when audio
data has been lost.