Fixed a failure to free context resulting in a segfault
Prior to this change, on my machine, after about 30 seconds or so of running the stream example, the program would segfault.
Upon investigation, I found that the segfaulting was happening because when the program calls state->mel_calc->calculate in
whisper_pcm_to_mel_with_state calls whisper_mel_init, which calls ggml_init, which uses a context from the global state.
However, that global context is never freed. This change frees that context and fixes the segfault, but I don't understand
the code enough to say whether it's the right way to fix it.