ci : add nvcc compilation cache to windows-cublas
This commit adds a directory to store the cuda compilation cache and
also caches this directory in github actions.

The motivation is to reduce the compilation time of the CUDA code. The
ccache should still help with the host part of the compilation but it
did not help with the CUDA code.

This also addes the --parallel (-j) flag to the cmake build command to
use all the available cores.