whisper.android.java: Add GGML_USE_CPU to whisper library
This commit adds GGML_USE_CPU to whisper library in order to enable the
CPU backend to work.

I'm not currently sure about the aproach to take with regards to the
backends on Android. I went down the dynamic library route, but it was
a struggle and perhaps not the best way to go. I'm still looking how to
enable GPU backends and which ones are available.